Adding Identity

+++ KemerburgazPetShop.Entities/Product.cs
@@ -13,7 +13,7 @@ namespace KemerburgazPetShop.Entities
         public string ProductName { get; set; }
         public string ProductDescription { get; set; }
         public string ImageURL { get; set; }
-        public decimal? ProductPrice { get; set; }
+        public double? ProductPrice { get; set; }
         public List<ProductCategory> ProductCategories { get; set; }
     }
 }
